feat: update readme, add jsdoc for generateMnemonic, rename chinese wordlist files, remove utils/download logic
