+++ src/cjs/wordlists/chineseSimplified.d.ts
@@ -1,0 +1,2 @@
+export declare const chineseSimplified: string[];
+//# sourceMappingURL=chineseSimplified.d.ts.map
