+++ src/cjs/wordlists/chineseTraditional.d.ts
@@ -1,0 +1,2 @@
+export declare const chineseTraditional: string[];
+//# sourceMappingURL=chineseTraditional.d.ts.map
