+++ ts_src/wordlists/chineseSimplified.ts
