--- types/wordlists.d.ts
@@ -1,5 +1,0 @@
-declare const wordlists: {
-    [index: string]: string[];
-};
-declare let _default: string[] | undefined;
-export { wordlists, _default };
